Save system & upgrade
Save system is now in place and working after having to start from scratch again as the old system just introduced more problems instead.

Added encryption and hashing, although not the strongest version right now, in the future will have to be made better.

made sure the first clicker upgrade is functional and in save system

second clicker upgrade is a WIP

Still a lot to do :(